swarm: move race/norace helpers from stream to testutil
As we will need to use the flag in other packages, too.
